hkd32: Factor `tiny-bip39` code into `hkd32`
Merges the `tiny-bip39` code `hkd32` properly, testing it against the
24-word BIP39 test vectors.
