+++ hkd32/Cargo.toml
@@ -23,8 +23,9 @@ travis-ci = { repository = "iqlusioninc/crates", branch = "develop" }
 [dependencies]
 getrandom = { version = "0.1", optional = true }
 hmac = { version = "0.7", default-features = false }
+lazy_static = { version = "1", optional = true, default-features = false }
 sha2 = { version = "0.8", default-features = false }
-tiny-bip39 = { version = "0.6", default-features = false, optional = true }
+pbkdf2 = { version = "0.3", optional = true, default-features = false }
 
 [dependencies.subtle-encoding]
 version = "0.4"
@@ -41,7 +42,7 @@ path = "../zeroize"
 default = ["alloc", "bech32", "getrandom"]
 alloc = ["zeroize/alloc"]
 bech32 = ["alloc", "subtle-encoding/bech32-preview"]
-mnemonic = ["alloc", "tiny-bip39"]
+mnemonic = ["alloc", "getrandom", "lazy_static", "pbkdf2"]
 
 [package.metadata.docs.rs]
 all-features = true
