+++ hkd32/src/lib.rs
@@ -41,7 +41,7 @@
 #![doc(html_root_url = "https://docs.rs/hkd32/0.1.2")]
 
 #[cfg(feature = "alloc")]
-#[cfg_attr(test, macro_use)]
+#[cfg_attr(any(feature = "mnemonic", test), macro_use)]
 extern crate alloc;
 
 mod key_material;
